// frontend/web/src/services/API/proserve-wb-provisioning-api/models/ProductParameter.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Product Provisioning API
 * API in charge of provisioning VEW products
 *
 * The version of the OpenAPI document: 2024-08-12T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { ParameterConstraints } from './ParameterConstraints';
import {
    ParameterConstraintsFromJSON,
    ParameterConstraintsFromJSONTyped,
    ParameterConstraintsToJSON,
} from './ParameterConstraints';
import type { ProductParameterMetaData } from './ProductParameterMetaData';
import {
    ProductParameterMetaDataFromJSON,
    ProductParameterMetaDataFromJSONTyped,
    ProductParameterMetaDataToJSON,
} from './ProductParameterMetaData';

/**
 * 
 * @export
 * @interface ProductParameter
 */
export interface ProductParameter {
    /**
     * Parameter key.
     * @type {string}
     * @memberof ProductParameter
     */
    parameterKey: string;
    /**
     * Default value.
     * @type {string}
     * @memberof ProductParameter
     */
    defaultValue?: string;
    /**
     * The description of the parameter.
     * @type {string}
     * @memberof ProductParameter
     */
    description?: string;
    /**
     * If this value is true, the value for this parameter is obfuscated from view when the parameter is retrieved.
     * @type {boolean}
     * @memberof ProductParameter
     */
    isNoEcho?: boolean;
    /**
     * The parameter type.
     * @type {string}
     * @memberof ProductParameter
     */
    parameterType?: string;
    /**
     * 
     * @type {ParameterConstraints}
     * @memberof ProductParameter
     */
    parameterConstraints?: ParameterConstraints;
    /**
     * 
     * @type {ProductParameterMetaData}
     * @memberof ProductParameter
     */
    parameterMetaData?: ProductParameterMetaData;
}

/**
 * Check if a given object implements the ProductParameter interface.
 */
export function instanceOfProductParameter(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "parameterKey" in value;

    return isInstance;
}

export function ProductParameterFromJSON(json: any): ProductParameter {
    return ProductParameterFromJSONTyped(json, false);
}

export function ProductParameterFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProductParameter {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'parameterKey': json['parameterKey'],
        'defaultValue': !exists(json, 'defaultValue') ? undefined : json['defaultValue'],
        'description': !exists(json, 'description') ? undefined : json['description'],
        'isNoEcho': !exists(json, 'isNoEcho') ? undefined : json['isNoEcho'],
        'parameterType': !exists(json, 'parameterType') ? undefined : json['parameterType'],
        'parameterConstraints': !exists(json, 'parameterConstraints') ? undefined : ParameterConstraintsFromJSON(json['parameterConstraints']),
        'parameterMetaData': !exists(json, 'parameterMetaData') ? undefined : ProductParameterMetaDataFromJSON(json['parameterMetaData']),
    };
}

export function ProductParameterToJSON(value?: ProductParameter | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'parameterKey': value.parameterKey,
        'defaultValue': value.defaultValue,
        'description': value.description,
        'isNoEcho': value.isNoEcho,
        'parameterType': value.parameterType,
        'parameterConstraints': ParameterConstraintsToJSON(value.parameterConstraints),
        'parameterMetaData': ProductParameterMetaDataToJSON(value.parameterMetaData),
    };
}